fix line-join tests
line-join/miter
I'm not sure where this changed but the output in v8 matches the output
in master. It seems v8 is more sensitive to differences.

line-join/round tests changed because roundLimit changed in the v8 spec.
It now uses miter joins instead of round joins where the angle is too
shallow for a round join to be visibly different.